Add a prefix for robot-sent messages
This is both to make the chat more easily interpretable, and to prevent things like impersonation.

The messages use `[]` instead of `<>` to make them quickly distinguishable from player messages.

The idea is that if `<>` were used, every time a robot sent a message, it would seem like a player sent that message, and players would have to read the username to see that it was a robot. Using `[]` should make it so a quick glance is enough to make clear the message isn't from a player, instead of making users read the username for every new could-be-player message.

In short it makes it clear and at-a-glance distinguishable which messages are human and which are from a robot, preventing possible frustrations and UX issues that would come from muddying up the chat with either the lack of a prefix, or the player-chat prefix, while still keeping it simple.