Increment version to 0.2.4, add new commands, fix debugging, add Mineclonia support, prefix robot messages (#2)
* Add Mineclonia crafting support

Might be better for wider support to check for the mcl_core mods presence instead of game name.

* Add mcl_core as optional dependency

* Fix debugging

Previously, when an error originated from a loaded file, you wouldn't see the error details. Asserting the files when loaded raises any errors they might raise so they can be properly debugged

* Inform user how to break non-empty robot

Previously, the user gets no indication of why their robot can't be broken when it has items.

* Add code-editing commands

* Add "regex" & "contains" condition commands

* Add a prefix for robot-sent messages

This is both to make the chat more easily interpretable, and to prevent things like impersonation.

The messages use `[]` instead of `<>` to make them quickly distinguishable from player messages.

The idea is that if `<>` were used, every time a robot sent a message, it would seem like a player sent that message, and players would have to read the username to see that it was a robot. Using `[]` should make it so a quick glance is enough to make clear the message isn't from a player, instead of making users read the username for every new could-be-player message.

In short it makes it clear and at-a-glance distinguishable which messages are human and which are from a robot, preventing possible frustrations and UX issues that would come from muddying up the chat with either the lack of a prefix, or the player-chat prefix, while still keeping it simple.

* Increment version to 0.2.4

* Minor whitespace changes

* Change "Loop" description to "Loop while"

This to indicate what kind of a loop it is and how to use it.

* (Minor) Fix a line's indentation

It used spaces instead of tabs

* Remove empty `depends` line